==> notebooks/0_exploratory_analysis_mssis_vessels.ipynb <==
# Exploratory Analysis Bermuda Case Study MSSIS Vessel History

Exploration of the Bermuda Case Study MSSIS vessel history dateset provided by Volpe.

file = '../data/raw/volpe-bermuda-mssis-vessel-history.csv'

out_png_path = '../img/0_exploratory_analysis_mssis_vessels_'

out_html_path = '../files/0_exploratory_analysis_mssis_vessels_'

## volpe-bermuda-mssis-vessel-history.csv

import pandas as pd
import geopandas
import folium
import matplotlib.pyplot as plt
import random

from shapely.geometry import Point

!head '-n 5' $file

df = pd.read_csv(file, sep=',',)

df.head()

df.info()

df.shape

df.describe()

df.isna().any()

df.EEZ.unique()

df.AO.unique()

df.SourceNum.unique()

df.RxStnId.unique()

df.EEZ.fillna('NONE', inplace=True)

df['TimeOfFix'] = pd.to_datetime(df['TimeOfFix'], unit='s')

df.head()

### GeoPandas and Folium

https://geopandas.org/gallery/plotting_with_folium.html  
https://jingwen-z.github.io/how-to-draw-a-variety-of-maps-with-folium-in-python/  

geometry = geopandas.points_from_xy(df.Longitude, df.Latitude)

geo_df = geopandas.GeoDataFrame(df, geometry=geometry)

geo_df.drop(columns=['Longitude', 'Latitude', 'SourceNum', 'RxStnId'], inplace=True)

geo_df.head()

# geo_df[['MMSI', 'PortIndex', 'AO']] = geo_df[['MMSI', 'PortIndex', 'AO']].astype('int32')
geo_df[['MMSI', 'PortIndex', 'AO']] = geo_df[['MMSI', 'PortIndex', 'AO']].astype('str')

geo_df.info()

world = geopandas.read_file(geopandas.datasets.get_path('naturalearth_lowres'))

fig, ax = plt.subplots(figsize=(24,18))
world.plot(ax=ax, alpha=0.4, color='grey')
geo_df.plot(column='MMSI', ax=ax, legend=True)
plt.title('MSSIS Bermuda Case Study - AIS Message Coverage for Vessels by MMSI')

fig.savefig(out_png_path + 'ais_global_coverage.png')

we can see there are a few outlier points over land in africa and antarctica

coasts = geopandas.read_file('https://raw.githubusercontent.com/nvkelso/natural-earth-vector/master/geojson/ne_10m_coastline.geojson')

coasts.total_bounds

fig, ax = plt.subplots(figsize=(24,18))

box = [-98.8085959644,18.4709430525,23.1835915356,63.9333578462]

xlim = ([box[0],  box[2]])
ylim = ([box[1],  box[3]])

ax.set_xlim(xlim)
ax.set_ylim(ylim)

coasts.plot(ax=ax, alpha=0.4, color='grey')
geo_df.plot(column='MMSI', ax=ax, legend=True)
plt.title('MSSIS Bermuda Case Study - AIS Message Trans-Atlantic Coverage for Vessels by MMSI')

fig.savefig(out_png_path + 'ais_transatlantic_coverage.png')

fig, ax = plt.subplots(figsize=(24,18))

box = [-53.4155724407,34.2808970679,13.9344472742,60.3598058554]

xlim = ([box[0],  box[2]])
ylim = ([box[1],  box[3]])

ax.set_xlim(xlim)
ax.set_ylim(ylim)


coasts.plot(ax=ax, alpha=0.4, color='grey')
geo_df.plot(column='MMSI', ax=ax, legend=True)
plt.title('MSSIS Bermuda Case Study - AIS Message North Atlantic Coverage for Vessels by MMSI')

fig.savefig(out_png_path + 'ais_north_atlantic_coverage.png')

fig, ax = plt.subplots(figsize=(24,18))

box = [-82.4230613878,24.2375385728,-59.6259944637,43.2329888635]

xlim = ([box[0],  box[2]])
ylim = ([box[1],  box[3]])

ax.set_xlim(xlim)
ax.set_ylim(ylim)


coasts.plot(ax=ax, alpha=0.4, color='grey')
geo_df.plot(column='MMSI', ax=ax, legend=True)
plt.title('MSSIS Bermuda Case Study - AIS Message U.S. Mid & South Atlantic Coverage for Vessels by MMSI')

fig.savefig(out_png_path + 'ais_us_mid_south_atlantic_coverage.png')

fig, ax = plt.subplots(figsize=(24,18))

box = [-68.6324891034,29.4053793797,-60.9058776912,35.1335421742]

xlim = ([box[0],  box[2]])
ylim = ([box[1],  box[3]])

ax.set_xlim(xlim)
ax.set_ylim(ylim)


coasts.plot(ax=ax, alpha=0.4, color='grey')
geo_df.plot(column='MMSI', ax=ax, legend=True)
plt.title('MSSIS Bermuda Case Study - AIS Message Bermuda Coverage for Vessels by MMSI')

fig.savefig(out_png_path + 'ais_bermuda_coverage.png')

### Vessels Calling Bermuda

bm_mmsi = ['244820000', '245227000', '538008204']

bm_df = geo_df[geo_df['MMSI'].isin(bm_mmsi)]

bm_df

fig, ax = plt.subplots(figsize=(24,18))

box = [-82.4230613878,24.2375385728,-59.6259944637,43.2329888635]

xlim = ([box[0],  box[2]])
ylim = ([box[1],  box[3]])

ax.set_xlim(xlim)
ax.set_ylim(ylim)


coasts.plot(ax=ax, alpha=0.4, color='grey')
bm_df.plot(column='MMSI', ax=ax, legend=True)
plt.title('MSSIS Bermuda Case Study - AIS Message Bermuda Trade Lane Coverage for Vessels by MMSI')

fig.savefig(out_png_path + 'ais_bermuda_trade_lane_coverage.png')

fig, ax = plt.subplots(figsize=(24,18))

box = [-82.4230613878,24.2375385728,-59.6259944637,43.2329888635]

xlim = ([box[0],  box[2]])
ylim = ([box[1],  box[3]])

ax.set_xlim(xlim)
ax.set_ylim(ylim)
geo_df[geo_df['MMSI'].isin(bm_mmsi)]

coasts.plot(ax=ax, alpha=0.4, color='grey')
bm_df[bm_df['MMSI'] == '245227000'].plot(column='MMSI', ax=ax, legend=True)
plt.title('MSSIS Bermuda Case Study - AIS Message Coverage for MV Somers Isles, MMSI: 245227000')

fig.savefig(out_png_path + 'ais_mv_somers_isles_coverage.png')

fig, ax = plt.subplots(figsize=(24,18))

box = [-82.4230613878,24.2375385728,-59.6259944637,43.2329888635]

xlim = ([box[0],  box[2]])
ylim = ([box[1],  box[3]])

ax.set_xlim(xlim)
ax.set_ylim(ylim)
geo_df[geo_df['MMSI'].isin(bm_mmsi)]

coasts.plot(ax=ax, alpha=0.4, color='grey')
bm_df[bm_df['MMSI'] == '244820000'].plot(column='MMSI', ax=ax, legend=True)
plt.title('MSSIS Bermuda Case Study - AIS Message Coverage for MV Bermuda Islander, MMSI: 244820000')

fig.savefig(out_png_path + 'ais_mv_bermuda_islander_coverage.png')

fig, ax = plt.subplots(figsize=(24,18))

box = [-82.4230613878,24.2375385728,-59.6259944637,43.2329888635]

xlim = ([box[0],  box[2]])
ylim = ([box[1],  box[3]])

ax.set_xlim(xlim)
ax.set_ylim(ylim)
geo_df[geo_df['MMSI'].isin(bm_mmsi)]

coasts.plot(ax=ax, alpha=0.4, color='grey')
bm_df[bm_df['MMSI'] == '538008204'].plot(column='MMSI', ax=ax, legend=True)
plt.title('MSSIS Bermuda Case Study - AIS Message Coverage for MV Oleander, MMSI: 538008204')

fig.savefig(out_png_path + 'ais_mv_oleander_coverage.png')

# https://stackoverflow.com/questions/28999287/generate-random-colors-rgb

number_of_colors = geo_df.MMSI.unique().size

color = ["#"+''.join([random.choice('0123456789ABCDEF') for j in range(6)])
             for i in range(number_of_colors)]

map = folium.Map(
    location = [32.316667,-64.833333],
    tiles='https://server.arcgisonline.com/ArcGIS/rest/services/World_Topo_Map/MapServer/tile/{z}/{y}/{x}',
    attr='Tiles &copy; Esri &mdash; EsriJapan, METI, Esri China (Hong Kong), and the GIS User Community',
    zoom_start=6
)

color_dict = {}

mmsi_list = geo_df.MMSI.unique()

# https://jingwen-z.github.io/how-to-draw-a-variety-of-maps-with-folium-in-python/
i=0
for mmsi in mmsi_list:
    color_dict[mmsi] = {
        'MMSI': mmsi,
        'color': color[i],
        'feature': folium.FeatureGroup(
                      name=mmsi,
                      show=False
                  )
    }
    i+=1

for idx, row in geo_df.iterrows():
    mmsi = row['MMSI']
    c = color_dict[mmsi]['color']

    folium.CircleMarker(
        location=[row.geometry.y, row.geometry.x],
        radius=1,
        color=c,
        fill_color=c,
        fill=True
    ).add_to(color_dict[mmsi]['feature'])

for k, v in color_dict.items():
    v['feature'].add_to(map)

folium.LayerControl(
    name='MMSI',
    collapsed=True
).add_to(map)

map.save(out_html_path + 'folium_map.html')

map
