==> notebooks/04_retrieve_nga_wpi.ipynb <==
# Retrieve NGA WPI 

import pandas as pd
import requests
import geopandas
import seaconex
import json

from shapely.geometry import Point
from geopandas import GeoDataFrame, read_file
Retrieve MSSIS ports from [Pub. 150, The World Port Index, National Geospatial-Intelligence Agency (NGA), Maritime Safety Office](https://msi.nga.mil/Publications/WPI) 

- [NGA Maritime Safety Information REST API Documentation](https://msi.nga.mil/api/swagger-ui.html)



import pandas as pd
import requests
import geopandas
import seaconex
import json
import numpy as np

from shapely.geometry import Point
from geopandas import GeoDataFrame, read_file

mssis_ais_region = '../data/interim/mssis-ais-region_v2.json'
mssis_port_visit_ports = '../data/interim/mssis-port-visit-ports_v1.json'
mssis_wpi_bm_ports = '../data/interim/mssis-wpi-bm-ports_v1.json'

df_ais = pd.read_json(mssis_ais_region)

df_port = pd.read_json(mssis_port_visit_ports)

df_bm = pd.read_json(mssis_wpi_bm_ports)

df_ais.head()

df_port.head()

df_bm.head()

df_bm = df_bm.filter(like='wpi').drop(columns=['mssis_wpi_region_index'])

# result = pd.concat([df_ais, df_port, df_bm], ignore_index=True, sort=True)

# result.head()

# result.filter(like='mssis').head()

# df_port.append(df_bm, ignore_index=True)

# merged_inner = pd.merge(left=df_port, right=df_bm, left_on='mssis_wpi', right_on='mssis_wpi')

# merged_inner

# join_col = df_port.columns & df_bm.columns

join_col = np.intersect1d(df_port.columns, df_bm.columns).tolist()

join_col

df = pd.merge(left=df_port, right=df_bm, how='outer', left_on=join_col, right_on=join_col)

join_col = np.intersect1d(df.columns, df_ais.columns).tolist()

join_col

result = pd.merge(left=df, right=df_ais, how='outer', left_on=join_col, right_on=join_col)

# result

result.info()

result.head()

result[['ais_count']] = result[['ais_count']].fillna(0).astype('int')

result

# result[['mssis_wpi_region_index', 'mssis_ao', 'ais_count']] = result[['mssis_wpi_region_index', 'mssis_ao', 'ais_count']].fillna(0).astype('int')

# result

result.sort_values('mssis_wpi').sort_values('ais_count')

wpi_list = result[result['mssis_wpi'] > 0].sort_values('mssis_wpi').mssis_wpi.unique().tolist()

wpi_list

nga_wpi = seaconex.nga_msi_wpi_api_get_world_port_index(wpi_list)

nga = pd.DataFrame(nga_wpi['ports'])

nga.to_json('../data/raw/' + 'nga-bermuda-wpi_v1.json')

nga.to_json('../data/raw/' + 'nga-bermuda-wpi_records_v1.json', orient='records')

nga.to_json('../data/raw/' + 'nga-bermuda-wpi_split_v1.json', orient='split')

nga.to_json('../data/raw/' + 'nga-bermuda-wpi_table_v1.json', orient='table')

# nga_wpi_prc = seaconex.nga_msi_wpi_api_get_prc_names()

# nga_wpi_prc

nga.head()

geometry = geopandas.points_from_xy(nga.xcoord, nga.ycoord)

nga.drop(columns=['ycoord', 'xcoord', 'latitude', 'longitude'], inplace=True)

null_columns=nga.columns[nga.isnull().all()]

null_columns

null_columns.tolist()

nga.drop(columns=null_columns, inplace=True)

print(nga.columns.tolist())

nga.columns.tolist()

pd.set_option("max_columns", None)
pd.set_option("max_rows", None)

drop_col = ['shelter', 'erTide', 'erSwell', 'erIce', 'erOther', 'overheadLimits', 'chDepth','anDepth','cpDepth','otDepth','tide', 'goodHoldingGround', 'turningArea', 'firstPortOfEntry', 'usRep', 'ptCompulsory', 'ptAdvisable', 'tugsSalvage', 'tugsAssist', 'qtPratique', 'qtOther', 'cmTelephone', 'cmTelegraph', 'cmRadio', 'cmRadioTel', 'cmAir', 'cmRail', 'loWharves', 'loAnchor', 'loMedMoor', 'loBeachMoor', 'loIceMoor', 'medFacilities', 'garbageDisposal', 'degauss', 'dirtyBallast', 'crFixed', 'crMobile', 'crFloating', 'lifts100', 'lifts50', 'lifts25', 'lifts0', 'srLongshore', 'srElectrical', 'srSteam', 'srNavigEquip', 'srElectRepair', 'suProvisions', 'suWater', 'suFuel', 'suDiesel', 'suDeck', 'repairCode', 'drydock', 'railway', 'qtSanitation', 'suAviationFuel', 'harborUse', 'ukcMgmtSystem', 'portSecurity', 'etaMessage', 'searchAndRescue', 'tss', 'vts', 'cht', 'loRoro', 'loSolidBulk', 'loContainer', 'loBreakBulk', 'loOilTerm', 'loLongTerm', 'loOther', 'loDangCargo', 'loLiquidBulk', 'srIceBreaking', 'srDiving', 'cranesContainer', 's121WaterBody', 's101Enc', 'dodWaterBody']

nga.drop(columns=drop_col, inplace=True)

nga

gdf = geopandas.GeoDataFrame(nga, geometry=geometry, crs="EPSG:4326")

gdf.head()

gdf.crs

gdf.to_file('../data/interim/' + 'nga-wpi_v1.json', driver='GeoJSON')

gdf.to_file('../data/interim/' + 'nga-wpi_v1.gpkg', driver='GPKG')

# are these ports missing??
salem = '8025'
cork = '34360'
london = '31470'
