==> notebooks/07_transport_calls.ipynb <==
import seaconex

import fiona

import pandas as pd
import numpy as np
import geopandas as gpd

from shapely.geometry import Point, LineString

seaconex.tools.edges.build_all_master_schedules_edges(
    '../data/gpkg/terminals.gpkg',
    '../data/raw/master-schedules-terminal-call-svc-proforma.csv',
    '../data/processed/master-schedules-terminal-call-svc-proforma.geojson',
    '../data/gpkg/master-schedules-terminal-call-svc-proforma.gpkg',
    '../data/processed/master-schedules-terminal-call-edge.geojson',
    '../data/gpkg/master-schedules-terminal-call-edge.gpkg',
    '../data/searoute/data/in/searoutes.csv'
)
