==> notebooks/08_generate_callback_json.ipynb <==
terminal_location = "../data/processed/terminals.geojson"
carrier_location = "../data/processed/carriers.json"
master_schedules_edges_location = "../data/processed/master-schedules-terminal-call-edge.geojson"
vessel_location = "../data/processed/vessels.json"

output_location = '../data/processed/lookups.json'

import seaconex

seaconex.build_all_lookups(
    terminal_location,
    carrier_location,
    master_schedules_edges_location,
    vessel_location,output_location
)
